Move code to __init__
